us002 merged to main

+++ src/main/java/com/bootcamp/be_java_hisp_w25_g02/dto/response/FollowingPostDTO.java
@@ -3,7 +3,7 @@ package com.bootcamp.be_java_hisp_w25_g02.dto.response;
 import java.util.List;
 
 public record FollowingPostDTO(
-    Long user_id,
+    Integer user_id,
     List<PostDTO> posts
 ) {
 }
